src/requirements.txt
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
datetime-truncate==1.1.1
discord-webhook==1.3.1
Django==5.0.7
django-bower==5.2.0
django-debug-toolbar==4.4.6
django-memoize==2.3.1
django-multiselectfield==0.1.13
filelock==3.15.4
frozenlist==1.4.1
gunicorn==22.0.0
huey==2.5.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
msgpack==1.0.8
packaging==24.1
peewee==3.17.6
pillow==10.4.0
protobuf==5.27.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-nvd3==0.16.0
python-slugify==8.0.4
PyYAML==6.0.1
ray==2.33.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
six==1.16.0
sqids==0.4.1
sqlparse==0.5.0
text-unidecode==1.3
urllib3==2.2.2
